requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<0.28
Pillow>=11.0,<12.0
python-dotenv>=1.0,<2.0
psycopg[binary]>=3.2,<4.0
django-filter>=24.0
django-redis>=5.4,<6.0
gunicorn>=22.0
whitenoise[brotli]>=6.7
